Add files via upload
This was my internship first Task offered  by Prodigy Info Tech in the field of Web Development. The task was to create a responsive landing page with an interactive navigation menu that changes colors or style when scrolled over or while hovering over menu item. The navigation menu is fixed and visible on all page. Using HTML to structure it, CSS to style the web page and JavaScript to add the interactivity for the web application. #WebDevelopment  #Internship  #ProdigyInfoTech.